python/pyproject.toml
[project]
name = "zero-to-AI"
version = "0.1.0"

# Note: this list of dependencies is intentionally larger than necessary
# as some of these aren't used in the zero-to-AI project.  I wanted to
# make the audience aware of these, and increase the size of the pypi_libs
# dataset and related graph. cj

dependencies = [
  "agent-framework",
  "aiohttp>=3.11.18",
  "alembic>=1.17.0",
  "av==16.0.1",
  "azure-ai-agents>=1.2.0b3",
  "azure-ai-documentintelligence==1.0.2",
  "azure-ai-evaluation",
  "azure-ai-projects",
  "azure-ai-textanalytics",
  "azure-core-tracing-opentelemetry",
  "azure-core==1.36.0",
  "azure-cosmos>=4.14.5",
  "azure-identity",
  "azure-keyvault-secrets>=4.10.0",
  "azure-mgmt-applicationinsights",
  "azure-mgmt-cognitiveservices",
  "azure-search-documents==11.5.2",
  "azure-storage-blob>=12.27.0",
  "beautifulsoup4>=4.14.2",
  "docopt",
  "duckdb>=1.4.3",
  "Faker",
  "fastapi[standard]",
  "fastmcp>=2.13.0",
  "geopy",
  "h11>=0.14.0",
  "httpx",
  "httpx[cli]>=0.28.1",
  "Jinja2>= 3.1.4",
  "jupyter>=1.1.1",
  "m26>=0.3.2",
  "markdown>=3.1.0",
  "matplotlib>=3.10.0",
  "openai>=1.96.0",
  "opentelemetry-api==1.38.0",
  "opentelemetry-exporter-otlp-proto-grpc==1.38.0",
  "opentelemetry-instrumentation-httpx==0.59b0",
  "opentelemetry-instrumentation-logging==0.59b0",
  "opentelemetry-instrumentation-psycopg2==0.59b0",
  "opentelemetry-instrumentation-requests==0.59b0",
  "opentelemetry-sdk==1.38.0",
  "pandas>=2.3.0", 
  "pgvector>=0.4.1",
  "psutil>=7.1.1",
  "psycopg2-binary>=2.9.11",
  "pydantic-core>=2.41.0",
  "pydantic>=2.12.0",
  "pylint>=4.0.4",
  "pytest-asyncio>=1.0.0",
  "pytest-cov>=6.1.1",
  "pytest-randomly>=4.0.0",
  "pytest>= 8.3.5",
  "python-dotenv>=1.2.1",
  "python-json-logger>=4.0.0",
  "python-multipart>=0.0.20",
  "pytz",
  "six==1.17.0",
  "sqlalchemy_utils",
  "SQLAlchemy>=2.0.44",
  "streamlit>=1.52.2",
  "tenacity>=9.1.2",
  "tiktoken>=0.12.0",
  "uvicorn==0.38.0",
  "watchdog"
]

requires-python = ">=3.13"
authors = [
  {name = "Chris Joakim", email = "chris.joakim@3cloudsolutions.com"}
]
maintainers = [
  {name = "Chris Joakim", email = "chris.joakim@3cloudsolutions.com"}
]
description = "The 'zero-to-AI' series of lessons at 3Cloud/Cognizant"
readme = "readme.md"
license = "MIT"
license-files = ["LICEN[CS]E.*"]
keywords = ["python", "AI", "3Cloud/Cognizant", "internal", "training"]
classifiers = [
  "Programming Language :: Python"
]

[project.urls]
Homepage = "https://github.com/cjoakim/zero-to-AI/"

# See https://docs.astral.sh/ruff/rules/
# ruff rule --all 

[tool.ruff]
line-length = 100

[tool.pylint]
max-line-length = 120          # Set maximum line length
output-format = "colorized"     # Set output format
